modulo_dos/smart-mobility-hub-old/backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2